+++ resources/lib/codex/modules/useI18n.cjs
@@ -1,0 +1,1 @@
+"use strict";const o=require("vue");function i(e,n,u=[]){const r=o.inject("CdxI18nFunction",void 0);return o.computed(()=>{const c=u.map(o.unref),t=r==null?void 0:r(e,...c);return t!=null?t:typeof n=="function"?n(...c):n})}module.exports=i;